KAFKA-4145; Avoid redundant integration testing in ProducerSendTests
Author: Vahid Hashemian <vahidhashemian@us.ibm.com>

Reviewers: Ismael Juma <ismael@juma.me.uk>

Closes #1842 from vahidhashemian/KAFKA-4145
